chore(tools): retire nwp/probe_nwp_model (mirrored by 'probe nwp --deep')
probe_nwp_model's per-backend availability dispatch now lives in the CLI
(_nwp_availability, reached via probe nwp --deep), so delete the tool and
its sole companion test (tests/nwp/test_tools.py). Port that test's two
cases — direct-https URL building + herbie-unavailable handling — onto
_nwp_availability in test_curate.py so the logic stays covered.
